Fix DST start/end month.  Calculate DST from ordinal week number.  Corrected WEEK_FORTH
day_from_week_and_weekday() is adequate for finding the last weekday of the month,
which is applicable to half the countries that use DST, but it does not work for
countries that change to/from DST on a day other than the last weekday of the month.
This commit introduces weekday_in_month() which works in all cases.  It does so by
more closely following DST start/end specifications, which do not refer to weeks.